refactor: update app name
