+++ voice_transcribe/configs/stage.yaml
