+++ voice_transcribe/dtos/settings/server_settings_dto.py
