+++ voice_transcribe/dtos/transcriber/transcriber_dto.py
@@ -1,6 +1,6 @@
 from pydantic import UUID4
 
-from whisper_transcribe.dtos.ws import WsRequestDto
+from voice_transcribe.dtos.ws import WsRequestDto
 
 
 class TranscriberDto(WsRequestDto):
