+++ voice_transcribe/routers.py
@@ -1,6 +1,6 @@
 from fastapi.routing import APIRouter
 
-from whisper_transcribe.controllers import TranscribeController, WsController
+from voice_transcribe.controllers import TranscribeController, WsController
 
 routes: list[APIRouter] = [
     TranscribeController().router,
