+++ voice_transcribe/shared/utils/funcs.py
